% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/swe_boundaries.R
\name{swe_boundaries}
\alias{swe_boundaries}
\title{Swe Boundaries}
\usage{
swe_boundaries(date, type = c("county", "municipal"), format = c("sp",
  "df"))
}
\arguments{
\item{date}{a date or a year}

\item{type}{type of unit, "county" or "municipal"}

\item{format}{format of return object, "df" for data.frame and "sp" for
SpatialPolygonsDataFrame}
}
\description{
Get swedish administrative boundaries for a specified year.
}

