% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{swe_counties}
\alias{swe_counties}
\title{Swedish county boundaries 1998-2007}
\format{A \code{SpatialPolygonsDataFrame} with 23 elements}
\source{
SCB \url{http://www.gis.scb.se/how/PXmap1/index.asp}
}
\usage{
data(swe_counties)
}
\description{
This data is a \code{SpatialPolygonsDataFrame} containing
  polygons of all counties in Sweden between 1998 and 2007.
}

