% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sweboundaries-pkg.R
\docType{package}
\name{sweboundaries}
\alias{sweboundaries}
\alias{sweboundaries-package}
\title{sweboundaries}
\description{
Swedish administrative boundaries for counties and
  municipalities from 1998 to 2008
}

